codecode

+++ Class 2/Free time folder 1-10/Freetime8.py
@@ -1,11 +1,11 @@
 print()
 print("Welcome to the bill calculator!")
 print()
-cost=int(input("Please enter the cost in pounds:"))
+cost=float(input("Please enter the cost in pounds:"))
 number=int(input("Please enter the number of people:"))
 Tip=(input("Do you want to add a tip(yes/no)?"))
 if Tip=="yes":
-    tip=int(input("How much do you want to add as a tip(in pounds)?"))
+    tip=float(input("How much do you want to add as a tip(in pounds)?"))
 
 else:
     tip=0
